+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Fixed
   * When filtering by game, the "Filtering by (...)" text would erroneously list an unrelated platform
   * Playtime graph would display timeline backwards
+  * Playtime graph with many dates would overlap (https://git.kucharczyk.xyz/lukas/timetracker/issues/34)
 
 ## 0.2.3 / 2023-01-15 23:13+01:00